Refactor importer into modular package
