+++ dropimator/__init__.py
@@ -1,0 +1,18 @@
+"""dropimator reusable Python package."""
+
+from .config import configure_logging, configure_openai, get_database_url
+from .csv_utils import find_csv_file, iter_csv_rows
+from .models import Base, Product
+from .product_service import enrich_products, process_csv_row
+
+__all__ = [
+    "Base",
+    "Product",
+    "configure_logging",
+    "configure_openai",
+    "enrich_products",
+    "find_csv_file",
+    "get_database_url",
+    "iter_csv_rows",
+    "process_csv_row",
+]
